% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pred.R
\name{as_class_pred}
\alias{as_class_pred}
\title{Coerce to a \code{class_pred} object}
\usage{
as_class_pred(x, which = integer(), equivocal = "[EQ]")
}
\arguments{
\item{x}{A factor or ordered factor.}

\item{which}{An integer vector specifying the locations of \code{x} to declare
as equivocal.}

\item{equivocal}{A single character specifying the equivocal label used
when printing.}
}
\description{
\code{as_class_pred()} provides coercion to \code{class_pred} from other
existing objects.
}
\examples{

x <- factor(c("Yes", "No", "Yes", "Yes"))
as_class_pred(x)

}
