% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boosting_predictions}
\alias{boosting_predictions}
\alias{boosting_predictions_oob}
\alias{boosting_predictions_test}
\title{Boosted regression trees predictions}
\value{
\item{boosting_predictions_oob,boosting_predictions_test}{tibbles}
}
\description{
Boosted regression trees predictions
}
\details{
These data have a set of holdout predictions from 10-fold
cross-validation and a separate collection of test set predictions from the
same boosted tree model. The data were generated using the \code{sim_regression}
function in the \pkg{modeldata} package.
}
\examples{
data(boosting_predictions_oob)
str(boosting_predictions_oob)
str(boosting_predictions_test)
}
\keyword{datasets}
