% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_prediction.R
\name{bound_prediction}
\alias{bound_prediction}
\title{Truncate a numeric prediction column}
\usage{
bound_prediction(x, lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{A data frame that contains a numeric column named \code{.pred}.}

\item{lower, upper}{Single numerics (or \code{NA}) that define constrains on \code{.pred}.}
}
\value{
\code{x} with potentially adjusted values.
}
\description{
For user-defined lower and/or upper bound, ensure that the values in the
\code{.pred} column are coerced to these bounds.
}
\examples{
data(solubility_test, package = "yardstick")

names(solubility_test) <- c("solubility", ".pred")

bound_prediction(solubility_test, lower = -1)
}
