% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-apply.R
\name{cal_apply}
\alias{cal_apply}
\title{Applies a calibration to a set of prediction probabilities}
\usage{
cal_apply(.data, object, prediction = NULL, threshold = NULL, ...)
}
\arguments{
\item{.data}{An object that can process a calibration object.}

\item{object}{The calibration object (\code{cal_object}).}

\item{prediction}{(Optional) Column identifier with the prediction.}

\item{threshold}{(Optional) Applies to binary models only. A number between
0 and 1. It contains the cut off for the prediction}

\item{...}{Optional arguments; currently unused.}
}
\description{
Applies a calibration to a set of prediction probabilities
}
\details{
It currently supports data.frames only. It extracts the \code{truth} and
the estimate columns names, and levels, from the calibration object.
}
\examples{
w_calibration <- cal_estimate_logistic(segment_logistic, Class)

cal_apply(segment_logistic, w_calibration)
}
