% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-estimate-isotonic.R
\name{cal_estimate_isotonic}
\alias{cal_estimate_isotonic}
\title{Uses an Isotonic regression model to calibrate probabilities}
\usage{
cal_estimate_isotonic(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  ...
)
}
\arguments{
\item{.data}{A \code{data.frame} object, or \code{tune_results} object, that contains
predictions and probability columns.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the class probabilities. It
defaults to the prefix used by tidymodels (\code{.pred_}). The order of the
identifiers will be considered the same as the order of the levels of the
\code{truth} variable.}

\item{...}{Additional arguments passed to the models or routines used to
calculate the new probabilities.}
}
\description{
Uses an Isotonic regression model to calibrate probabilities
}
\details{
This function uses \code{stats::isoreg()} to create obtain the calibration
values.
}
\examples{
# It will automatically identify the probability columns
# if passed a model fitted with tidymodels
cal_estimate_isotonic(segment_logistic, Class)
# Specify the variable names in a vector of unquoted names
cal_estimate_isotonic(segment_logistic, Class, c(.pred_poor, .pred_good))
# dplyr selector functions are also supported
cal_estimate_isotonic(segment_logistic, Class, dplyr::starts_with(".pred_"))
}
\references{
Zadrozny, Bianca and Elkan, Charles. (2002). Transforming Classifier Scores
into Accurate Multiclass Probability Estimates. \emph{Proceedings of the ACM SIGKDD
International Conference on Knowledge Discovery and Data Mining.}
}
