% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-estimate-multinom.R
\name{cal_estimate_multinomial}
\alias{cal_estimate_multinomial}
\alias{cal_estimate_multinomial.data.frame}
\alias{cal_estimate_multinomial.tune_results}
\title{Uses a Multinomial calibration model to calculate new probabilities}
\usage{
cal_estimate_multinomial(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  smooth = TRUE,
  parameters = NULL,
  ...
)

\method{cal_estimate_multinomial}{data.frame}(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  smooth = TRUE,
  parameters = NULL,
  ...
)

\method{cal_estimate_multinomial}{tune_results}(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  smooth = TRUE,
  parameters = NULL,
  ...
)
}
\arguments{
\item{.data}{A \code{data.frame} object, or \code{tune_results} object, that contains
predictions and probability columns.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the class probabilities. It
defaults to the prefix used by tidymodels (\code{.pred_}). The order of the
identifiers will be considered the same as the order of the levels of the
\code{truth} variable.}

\item{smooth}{Applies to the logistic models. It switches between logistic
spline when \code{TRUE}, and simple logistic regression when \code{FALSE}.}

\item{parameters}{(Optional)  An optional tibble of tuning parameter values
that can be used to filter the predicted values before processing. Applies
only to \code{tune_results} objects.}

\item{...}{Additional arguments passed to the models or routines used to
calculate the new probabilities.}
}
\description{
Uses a Multinomial calibration model to calculate new probabilities
}
\details{
It uses the \code{multinom} function, from the \code{nnet} package, to
create the calibration.
}
