% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{cal_isotonic_boot}
\alias{cal_isotonic_boot}
\title{Uses a bootstrapped Isotonic regression model to calibrate probabilities}
\usage{
cal_isotonic_boot(
  .data,
  truth = NULL,
  estimate = NULL,
  times = 10,
  event_level = c("first", "second"),
  ...
)
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{estimate}{The column identifier for the prediction probabilities.
This should be an unquoted column name}

\item{times}{Number of bootstraps.}

\item{event_level}{single string. Either "first" or "second" to specify which
level of truth to consider as the "event".}

\item{...}{Optional arguments; currently unused.}
}
\description{
Uses a bootstrapped Isotonic regression model to calibrate probabilities
}
