% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-plot.R
\name{cal_plot_compare_breaks}
\alias{cal_plot_compare_breaks}
\title{Compare one or several calibrations}
\usage{
cal_plot_compare_breaks(
  .data,
  ...,
  num_breaks = 10,
  .original_name = "Original"
)
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns.}

\item{...}{One or multiple \code{cal_objects}. They can be named.}

\item{num_breaks}{The number of segments to group the probabilities. It
defaults to 10.}

\item{.original_name}{Label for the original probabilities. It default to
"Original". Setting to \code{NULL} will prevent the original probabilities from
being added to this function's output.}
}
\description{
Compare one or several calibrations
}
