% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rank.R
\name{cal_rank}
\alias{cal_rank}
\alias{cal_rank.resample_results}
\title{Determine the best calibration method using resampling}
\usage{
cal_rank(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  metrics = NULL,
  ...
)

\method{cal_rank}{resample_results}(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  metrics = NULL,
  summarize = TRUE,
  level = 0.9,
  ...
)
}
\arguments{
\item{.data}{An \code{rset} object or the results of \code{\link[tune:fit_resamples]{tune::fit_resamples()}} with
a \code{.predictions} column.}

\item{truth}{The column identifier for the true results. This should be an
unquoted column name. The nature of the column depends on the problem type.
\itemize{
\item regression models: a single numeric column.
\item classification models: a single factor column.
}

For objects produced by \code{\link[tune:fit_resamples]{tune::fit_resamples()}} is argument is ignored.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the predictions.
\itemize{
\item regression models: a single numeric column.
\item classification models: multiple columns containing class probabilities. The
order of the identifiers will be considered the same as the order of the
levels of the \code{truth} variable.
}

For objects produced by \code{\link[tune:fit_resamples]{tune::fit_resamples()}} is argument is ignored.}

\item{metrics}{A set of metrics passed created via \code{yardstick::metric_set()}}

\item{...}{Other options passed}

\item{summarize}{Should the overall ranks be computed (the default) or should
the resample-level results be returned?}

\item{level}{Confidence level for two-sided intervals produced by
\code{\link[stats:t.test]{stats::t.test()}} (\code{summarize = TRUE} only)}
}
\description{
\code{cal_rank()} applies multiple calibration methods and results information on
the quality of the calibration relative to no calibration using resampling.
}
