% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pred.R
\name{class_pred}
\alias{class_pred}
\title{Create a class prediction object}
\usage{
class_pred(x = factor(), which = integer())
}
\arguments{
\item{x}{A factor or ordered factor.}

\item{which}{An integer vector specifying the locations of \code{x} to declare
as equivocal.}
}
\description{
\code{class_pred()} creates a \code{class_pred} object from a factor or ordered
factor. You can optionally specify values of the factor to be set
as \emph{equivocal}.
}
\details{
Equivocal values are those that you feel unsure about, and would like to
exclude from performance calculations or other metrics.

By default, \code{probably} prints equivocal values as \code{EQ}. You can change this
with the global option, \code{probably.equivocal_label}.
}
\examples{

x <- factor(c("Yes", "No", "Yes", "Yes"))

# Create a class_pred object from a factor
class_pred(x)

# Say you aren't sure about that 2nd "Yes" value. You could mark it as
# equivocal.
class_pred(x, which = 3)

}
