% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-validate.R
\name{collect_metrics.cal_rset}
\alias{collect_metrics.cal_rset}
\title{Obtain and format metrics produced by calibration validation}
\usage{
\method{collect_metrics}{cal_rset}(x, summarize = TRUE, ...)
}
\arguments{
\item{x}{An object produced by one of the validation function (or class
\code{cal_rset}).}

\item{summarize}{A logical; should metrics be summarized over resamples
(\code{TRUE}) or return the values for each individual resample. See
\code{\link[tune:collect_predictions]{tune::collect_metrics()}} for more details.}

\item{...}{Not currently used.}
}
\value{
A tibble
}
\description{
Obtain and format metrics produced by calibration validation
}
