% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-plot.R
\name{.cal_table_compare}
\alias{.cal_table_compare}
\title{Compare one or several calibrations}
\usage{
.cal_table_compare(.data, ..., .original_name = "Original")
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns.}

\item{...}{One or multiple \code{cal_objects}. They can be named.}

\item{.original_name}{Label for the original probabilities. It default to
"Original". Setting to \code{NULL} will prevent the original probabilities from
being added to this function's output.}
}
\description{
Compare one or several calibrations
}
\keyword{internal}
