% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butcher.R
\name{inf_conformal-butcher}
\alias{inf_conformal-butcher}
\alias{axe_call.int_conformal_full}
\alias{axe_ctrl.int_conformal_full}
\alias{axe_data.int_conformal_full}
\alias{axe_env.int_conformal_full}
\alias{axe_fitted.int_conformal_full}
\alias{axe_call.int_conformal_split}
\alias{axe_ctrl.int_conformal_split}
\alias{axe_data.int_conformal_split}
\alias{axe_env.int_conformal_split}
\alias{axe_fitted.int_conformal_split}
\alias{axe_call.int_conformal_quantile}
\alias{axe_ctrl.int_conformal_quantile}
\alias{axe_data.int_conformal_quantile}
\alias{axe_env.int_conformal_quantile}
\alias{axe_fitted.int_conformal_quantile}
\alias{axe_call.int_conformal_cv}
\alias{axe_ctrl.int_conformal_cv}
\alias{axe_data.int_conformal_cv}
\alias{axe_env.int_conformal_cv}
\alias{axe_fitted.int_conformal_cv}
\title{Butcher methods for conformal inteference intervals}
\usage{
axe_call.int_conformal_full(x, verbose = FALSE, ...)

axe_ctrl.int_conformal_full(x, verbose = FALSE, ...)

axe_data.int_conformal_full(x, verbose = FALSE, ...)

axe_env.int_conformal_full(x, verbose = FALSE, ...)

axe_fitted.int_conformal_full(x, verbose = FALSE, ...)

axe_call.int_conformal_split(x, verbose = FALSE, ...)

axe_ctrl.int_conformal_split(x, verbose = FALSE, ...)

axe_data.int_conformal_split(x, verbose = FALSE, ...)

axe_env.int_conformal_split(x, verbose = FALSE, ...)

axe_fitted.int_conformal_split(x, verbose = FALSE, ...)

axe_call.int_conformal_quantile(x, verbose = FALSE, ...)

axe_ctrl.int_conformal_quantile(x, verbose = FALSE, ...)

axe_data.int_conformal_quantile(x, verbose = FALSE, ...)

axe_env.int_conformal_quantile(x, verbose = FALSE, ...)

axe_fitted.int_conformal_quantile(x, verbose = FALSE, ...)

axe_call.int_conformal_cv(x, verbose = FALSE, ...)

axe_ctrl.int_conformal_cv(x, verbose = FALSE, ...)

axe_data.int_conformal_cv(x, verbose = FALSE, ...)

axe_env.int_conformal_cv(x, verbose = FALSE, ...)

axe_fitted.int_conformal_cv(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A conformal inference interval object.}

\item{verbose}{Should information be printed about how much memory is freed
from butchering?}

\item{...}{Extra arguments possibly used by underlying methods.}
}
\description{
These methods allow you to use the butcher package to reduce the size of a
conformal inference interval object. After calling \code{butcher::butcher()} on a
conformal inference interval object, the only guarantee is that you will
still be able to \code{predict()} from that conformal inference interval object.
Other functions may not work as expected.
}
