% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pred.R
\name{is_class_pred}
\alias{is_class_pred}
\title{Test if an object inherits from \code{class_pred}}
\usage{
is_class_pred(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\code{is_class_pred()} checks if an object is a \code{class_pred} object.
}
\examples{

x <- class_pred(factor(1:5))

is_class_pred(x)

}
