% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pred.R
\name{levels.class_pred}
\alias{levels.class_pred}
\title{Extract \code{class_pred} levels}
\usage{
\method{levels}{class_pred}(x)
}
\arguments{
\item{x}{A \code{class_pred} object.}
}
\description{
The levels of a \code{class_pred} object do \emph{not} include the equivocal value.
}
\examples{

x <- class_pred(factor(1:5), which = 1)

# notice that even though `1` is not in the `class_pred` vector, the
# level remains from the original factor
levels(x)

}
