% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_infer.R, R/conformal_infer_cv.R
\name{predict.int_conformal_infer}
\alias{predict.int_conformal_infer}
\alias{predict.int_conformal_infer_cv}
\title{Prediction intervals from conformal methods}
\usage{
\method{predict}{int_conformal_infer}(object, new_data, level = 0.95, ...)

\method{predict}{int_conformal_infer_cv}(object, new_data, level = 0.95, ...)
}
\arguments{
\item{object}{An object produced by \code{\link[=predict.int_conformal_infer]{predict.int_conformal_infer()}}.}

\item{new_data}{A data frame of predictors.}

\item{level}{The confidence level for the intervals.}

\item{...}{Not currently used.}
}
\value{
A tibble with columns \code{.pred_lower} and \code{.pred_upper}. If
the computations for the prediction bound fail, a missing value is used.
}
\description{
Prediction intervals from conformal methods
}
\seealso{
\code{\link[=int_conformal_infer]{int_conformal_infer()}}, \code{\link[=int_conformal_infer_cv]{int_conformal_infer_cv()}}
}
