% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probably-package.R
\docType{package}
\name{probably-package}
\alias{probably}
\alias{probably-package}
\title{probably: Tools for Post-Processing Class Probability Estimates}
\description{
Models can be improved by post-processing class probabilities, by: recalibration, conversion to hard probabilities, assessment of equivocal zones, and other activities. 'probably' contains tools for conducting these operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/probably/}
  \item \url{https://probably.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/probably/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com}

Authors:
\itemize{
  \item Davis Vaughan \email{davis@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
