% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/vctrs-compat.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{augment}
\alias{required_pkgs}
\alias{collect_metrics}
\alias{collect_predictions}
\alias{as.factor}
\alias{as.ordered}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:coercion-factor]{as.factor}}, \code{\link[generics:coercion-factor]{as.ordered}}, \code{\link[generics]{augment}}, \code{\link[generics]{fit}}, \code{\link[generics]{required_pkgs}}}

  \item{tune}{\code{\link[tune:collect_predictions]{collect_metrics}}, \code{\link[tune]{collect_predictions}}}
}}

