% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pred.R
\name{reportable_rate}
\alias{reportable_rate}
\title{Calculate the reportable rate}
\usage{
reportable_rate(x)
}
\arguments{
\item{x}{A \code{class_pred} object.}
}
\description{
The \emph{reportable rate} is defined as the percentage of class predictions
that are \emph{not} equivocal.
}
\details{
The reportable rate is calculated as \code{(n_not_equivocal / n)}.
}
\examples{

x <- class_pred(factor(1:5), which = c(1, 2))

# 3 / 5
reportable_rate(x)

}
