% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-estimate-beta.R, R/cal-estimate-linear.R,
%   R/cal-estimate-logistic.R, R/cal-estimate-multinom.R, R/cal-estimate-none.R,
%   R/cal-pkg-check.R
\name{required_pkgs.cal_estimate_beta}
\alias{required_pkgs.cal_estimate_beta}
\alias{required_pkgs.cal_estimate_linear_spline}
\alias{required_pkgs.cal_estimate_linear}
\alias{required_pkgs.cal_estimate_logistic_spline}
\alias{required_pkgs.cal_estimate_logistic}
\alias{required_pkgs.cal_estimate_multinomial_spline}
\alias{required_pkgs.cal_estimate_multinomial}
\alias{required_pkgs.cal_estimate_none}
\alias{required_pkgs.cal_object}
\title{S3 methods to track which additional packages are needed for specific
calibrations}
\usage{
\method{required_pkgs}{cal_estimate_beta}(x, ...)

\method{required_pkgs}{cal_estimate_linear_spline}(x, ...)

\method{required_pkgs}{cal_estimate_linear}(x, ...)

\method{required_pkgs}{cal_estimate_logistic_spline}(x, ...)

\method{required_pkgs}{cal_estimate_logistic}(x, ...)

\method{required_pkgs}{cal_estimate_multinomial_spline}(x, ...)

\method{required_pkgs}{cal_estimate_multinomial}(x, ...)

\method{required_pkgs}{cal_estimate_multinomial}(x, ...)

\method{required_pkgs}{cal_estimate_none}(x, ...)

\method{required_pkgs}{cal_object}(x, ...)
}
\arguments{
\item{x}{A calibration object}

\item{...}{Other arguments passed to methods}
}
\description{
S3 methods to track which additional packages are needed for specific
calibrations
}
\keyword{internal}
