% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_infer_cv.R, R/conformal_infer_full.R,
%   R/conformal_infer_quantile.R, R/conformal_infer_split.R
\name{required_pkgs.int_conformal_cv}
\alias{required_pkgs.int_conformal_cv}
\alias{required_pkgs.int_conformal_full}
\alias{required_pkgs.int_conformal_quantile}
\alias{required_pkgs.int_conformal_split}
\title{S3 methods to track which additional packages are needed for prediction
intervals via conformal inference}
\usage{
\method{required_pkgs}{int_conformal_cv}(x, infra = TRUE, ...)

\method{required_pkgs}{int_conformal_full}(x, infra = TRUE, ...)

\method{required_pkgs}{int_conformal_quantile}(x, infra = TRUE, ...)

\method{required_pkgs}{int_conformal_split}(x, infra = TRUE, ...)
}
\arguments{
\item{x}{a conformal interval object}

\item{...}{Other arguments passed to methods}
}
\description{
S3 methods to track which additional packages are needed for prediction
intervals via conformal inference
}
\keyword{internal}
