% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segment_naive_bayes}
\alias{segment_naive_bayes}
\alias{segment_logistic}
\title{Image segmentation predictions}
\source{
Hill, LaPan, Li and Haney (2007). Impact of image segmentation on
high-content screening data quality for SK-BR-3 cells, \emph{BMC
Bioinformatics}, Vol. 8, pg. 340,
\url{http://www.biomedcentral.com/1471-2105/8/340}.
}
\value{
\item{segment_naive_bayes,segment_logistic}{a tibble}
}
\description{
Image segmentation predictions
}
\details{
These objects contain test set predictions for the cell segmentation
data from Hill, LaPan, Li and Haney (2007). Each data frame are the results
from different models (naive Bayes and logistic regression).
}
\examples{
data(segment_naive_bayes)
data(segment_logistic)
}
\keyword{datasets}
