% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_data.R
\name{threshold_data}
\alias{threshold_data}
\alias{threshold_data.data.frame}
\title{Generate data for thresholding two-class data}
\usage{
threshold_data(.data, ...)

\method{threshold_data}{data.frame}(.data, truth, estimate,
  thresholds = NULL, na_rm = TRUE, ...)
}
\arguments{
\item{.data}{A tibble or data frame, potentially grouped.}

\item{...}{Currently unused.}

\item{truth}{A two-level factor of the true outcome labels.
Note that that the global option \code{yardstick.event_first} will be
used to determine which level is the event of interest.}

\item{estimate}{A numeric vector of class probabilities such
that larger values are more indicative of the event of interest.}

\item{thresholds}{A numeric vector of values for the probability
threshold to call a sample an event. If unspecified, a series
of values between 0.5 and 1.0 are used. \strong{Note}: if this
argument is used, it must be named.}

\item{na_rm}{A single logical: should missing data be removed?}
}
\value{
A tibble with columns for the performance
characteristics, the threshold, and existing groups.
}
\description{
\code{threshold_data} can take a set of class probability predictions
and determine performance characteristics across different values
of the probability threshold and any existing groups.
}
