% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\name{vec_cast.class_pred}
\alias{vec_cast.class_pred}
\title{Cast a \code{class_pred} vector to a specified type}
\usage{
\method{vec_cast}{class_pred}(x, to)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}
}
\description{
Cast a \code{class_pred} vector to a specified type
}
