% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\name{vec_proxy_equal.class_pred}
\alias{vec_proxy_equal.class_pred}
\title{Equality for \code{class_pred}}
\usage{
\method{vec_proxy_equal}{class_pred}(x)
}
\description{
\code{class_pred} objects are converted to integer before equality checks
are done.
}
\keyword{internal}
