% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\name{vec_ptype2.class_pred}
\alias{vec_ptype2.class_pred}
\title{Find the common type for a \code{class_pred} and another object}
\usage{
\method{vec_ptype2}{class_pred}(x, y, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Find the common type for a \code{class_pred} and another object
}
