% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\name{vec_type2.class_pred}
\alias{vec_type2.class_pred}
\title{Find the common type for a \code{class_pred} and another object}
\usage{
\method{vec_type2}{class_pred}(x, y)
}
\arguments{
\item{x}{Either vector types; i.e.}

\item{y}{Either vector types; i.e.}
}
\description{
Find the common type for a \code{class_pred} and another object
}
