% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-apply.R
\name{cal_apply}
\alias{cal_apply}
\title{Applies a calibration to a set of prediction probabilities}
\usage{
cal_apply(x, calibration, ...)
}
\arguments{
\item{x}{An object that can process a calibration object.}

\item{calibration}{The calibration object (\code{cal_object}).}

\item{...}{Optional arguments; currently unused.}
}
\description{
Applies a calibration to a set of prediction probabilities
}
\details{
It currently supports data.frames only. It extracts the \code{truth} and
the estimate columns names, and levels, from the calibration object.
}
