% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{cal_binary_plot_breaks}
\alias{cal_binary_plot_breaks}
\alias{cal_binary_plot_logistic}
\alias{cal_binary_plot_windowed}
\title{Probability Calibration plots}
\usage{
cal_binary_plot_breaks(
  .data,
  truth,
  estimate,
  num_breaks = 10,
  conf_level = 0.9,
  include_ribbon = TRUE,
  include_rug = TRUE
)

cal_binary_plot_logistic(
  .data,
  truth,
  estimate,
  conf_level = 0.9,
  include_rug = TRUE,
  include_ribbon = TRUE
)

cal_binary_plot_windowed(
  .data,
  truth,
  estimate,
  window_size = 100,
  step_size = window_size,
  conf_level = 0.9,
  include_ribbon = TRUE,
  include_rug = TRUE
)
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name}

\item{estimate}{The column identifier for the prediction probabilities.
This should be an unquoted column name}

\item{num_breaks}{The number of segments to group the probabilities. Defaults
to 10}

\item{conf_level}{Confidence level to use in the visualization. Defaults to 0.9}

\item{include_ribbon}{Flag that indicates if the ribbon layer is to be
included. Defaults to \code{TRUE}}

\item{include_rug}{Flag that indicates if the Rug layer is to be included.
Defaults to \code{TRUE}}

\item{window_size}{The size of segments. Used for the windowed probability
calculations}

\item{step_size}{The gap between segments. Used for the windowed probability
calculations}
}
\description{
Calibration plot functions. They require a data.frame that contains
the predictions and probability columns. The output is a \code{ggplot2} graph.
}
\examples{

library(ggplot2)

cal_binary_plot_breaks(
  segment_logistic,
  Class,
  .pred_good
)

cal_binary_plot_logistic(
  segment_logistic,
  Class,
  .pred_good
)

cal_binary_plot_windowed(
  segment_logistic,
  Class,
  .pred_good
)

}
