% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-estimate-beta.R
\name{cal_estimate_beta}
\alias{cal_estimate_beta}
\title{Uses a Beta calibration model to calculate new probabilities}
\usage{
cal_estimate_beta(
  .data,
  truth = NULL,
  shape_params = 2,
  location_params = 1,
  estimate = dplyr::starts_with(".pred_"),
  ...
)
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{shape_params}{Number of shape parameters to use. Accepted values are
1 and 2. Defaults to 2.}

\item{location_params}{Number of location parameters to use. Accepted values
1 and 0. Defaults to 1.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the class probabilities. It
defaults to the prefix used by tidymodels (\code{.pred_}). The order of the
identifiers will be considered the same as the order of the levels of the
\code{truth} variable.}

\item{...}{Additional arguments passed to the models or routines used to
calculate the new probabilities.}
}
\description{
Uses a Beta calibration model to calculate new probabilities
}
\details{
This function uses the \code{betcal::beta_calibration()} function, and
retains the resulting model.
}
\examples{
# It will automatically identify the probability columns
# if passed a model fitted with tidymodels
cal_estimate_beta(segment_logistic, Class)
}
\references{
Meelis Kull, Telmo M. Silva Filho, Peter Flach "Beyond sigmoids:
How to obtain well-calibrated probabilities from binary classifiers with beta
calibration," \emph{Electronic Journal of Statistics} 11(2), 5052-5080, (2017)
}
