% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-estimate.R
\name{cal_estimate_isotonic}
\alias{cal_estimate_isotonic}
\title{Uses an Isotonic regression model to calibrate probabilities}
\usage{
cal_estimate_isotonic(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  ...
)
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the class probabilities. It
defaults to the prefix used by tidymodels (\code{.pred_}). The order of the
identifiers will be considered the same as the order of the levels of the
\code{truth} variable.}

\item{...}{Additional arguments passed to the models or routines used to
calculate the new probabilities.}
}
\description{
Uses an Isotonic regression model to calibrate probabilities
}
