% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{cal_isotonic}
\alias{cal_isotonic}
\title{Uses an Isotonic regression model to calibrate probabilities}
\usage{
cal_isotonic(
  .data,
  truth = NULL,
  estimate = NULL,
  event_level = c("first", "second"),
  ...
)
}
\arguments{
\item{.data}{A data.frame object containing predictions and probability columns.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the class probabilities. It
defaults to the prefix used by tidymodels (\code{.pred_}). The order of the
identifiers will be considered the same as the order of the levels of the
\code{truth} variable.}

\item{event_level}{single string. Either "first" or "second" to specify which
level of truth to consider as the "event".}

\item{...}{Optional arguments; currently unused.}
}
\description{
Uses an Isotonic regression model to calibrate probabilities
}
