% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_rank.R
\name{cal_rank}
\alias{cal_rank}
\alias{cal_rank.resample_results}
\alias{cal_rank.rset}
\title{Determine the best calibration method using resampling}
\usage{
cal_rank(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  metrics = NULL,
  ...
)

\method{cal_rank}{resample_results}(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  metrics = NULL,
  summarize = TRUE,
  ...
)

\method{cal_rank}{rset}(
  .data,
  truth = NULL,
  estimate = dplyr::starts_with(".pred_"),
  metrics = NULL,
  summarize = TRUE,
  ...
)
}
\arguments{
\item{.data}{An \code{rset} object or the results of \code{\link[tune:fit_resamples]{tune::fit_resamples()}} with
a \code{.predictions} column.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should be an unquoted column name.}

\item{estimate}{A vector of column identifiers, or one of \code{dplyr} selector
functions to choose which variables contains the class probabilities. It
defaults to the prefix used by tidymodels (\code{.pred_}). The order of the
identifiers will be considered the same as the order of the levels of the
\code{truth} variable.}

\item{metrics}{A set of metrics passed created via \code{yardstick::metric_set()}}

\item{...}{Options to pass to \code{\link[=cal_estimate_logistic]{cal_estimate_logistic()}}, such as the \code{smooth}
argument.}

\item{summarize}{Should the overall ranks be computed (the default) or should
the resample-level results be returned?}
}
\description{
Determine the best calibration method using resampling
}
