% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-validate.R
\name{cal_validate_summarize}
\alias{cal_validate_summarize}
\alias{cal_validate_summarize.cal_rset}
\title{Summarizes the metrics of a Calibrated Re-sampled set}
\usage{
cal_validate_summarize(x)

\method{cal_validate_summarize}{cal_rset}(x)
}
\arguments{
\item{x}{Calibrated Re-sampled set}
}
\description{
Summarizes the metrics of a Calibrated Re-sampled set
}
