% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-validate.R
\name{cal_validate_summarize}
\alias{cal_validate_summarize}
\alias{cal_validate_summarize.cal_rset}
\title{Summarizes the metrics of a Calibrated Re-sampled set}
\usage{
cal_validate_summarize(x)

\method{cal_validate_summarize}{cal_rset}(x)
}
\arguments{
\item{x}{Calibrated Re-sampled set}
}
\description{
Summarizes the metrics of a Calibrated Re-sampled set
}
\examples{

library(magrittr)

sl_val <- segment_logistic \%>\%
  rsample::vfold_cv() \%>\%
  cal_validate_beta(Class, summarize = FALSE, save_details = TRUE)

sl_val

cal_validate_summarize(sl_val)

}
