% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probably-package.R
\docType{package}
\name{probably-package}
\alias{probably}
\alias{probably-package}
\title{probably: Tools for Post-Processing Class Probability Estimates}
\description{
Models can be improved by post-processing class probabilities, by: recalibration, conversion to hard probabilities, assessment of equivocal zones, and other activities. 'probably' contains tools for conducting these operations as well as calibration tools and conformal inference techniques for regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/probably}
  \item \url{https://probably.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/probably/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co}

Authors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co}
  \item Edgar Ruiz \email{edgar@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
