% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.factor}
\alias{as.ordered}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{as.factor}}, \code{\link[generics]{as.ordered}}}
}}

