% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compat.R
\name{vec_cast.class_pred}
\alias{vec_cast.class_pred}
\title{Cast a \code{class_pred} vector to a specified type}
\usage{
\method{vec_cast}{class_pred}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()} and \code{vec_restore()}, these dots are only for future
extensions and should be empty.}
}
\description{
Cast a \code{class_pred} vector to a specified type
}
